+++ icekit/plugins/image/abstract_models.py
@@ -1,12 +1,17 @@
+# -*- coding: utf-8 -*-
+
 from django.core.exceptions import ValidationError
 from django.template import Context
+from django.template.defaultfilters import filesizeformat
 from django.template.loader import render_to_string
+from django.utils import timezone
 from django.utils.safestring import mark_safe
 from django.db import models
 from django.utils import six
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
 from fluent_contents.models import ContentItem
+from icekit.fields import QuietImageField
 
 
 @python_2_unicode_compatible
@@ -14,14 +19,16 @@ class AbstractImage(models.Model):
     """
     A reusable image.
     """
-    image = models.ImageField(
+    image = QuietImageField(
         upload_to='uploads/images/',
         verbose_name=_('Image file'),
+        width_field="width",
+        height_field="height",
     )
     title = models.CharField(
         max_length=255,
         blank=True,
-        help_text=_('Can be included in captions'),
+        help_text=_('You must specify either title or help text. Title can be included in captions.'),
     )
     alt_text = models.CharField(
         max_length=255,
@@ -31,17 +38,52 @@ class AbstractImage(models.Model):
     caption = models.TextField(
         blank=True,
     )
+    credit = models.CharField(
+        max_length=255,
+        blank=True,
+        help_text=_("Who or what to credit whenever the image is used."),
+    )
+    source = models.CharField(
+        max_length=255,
+        blank=True,
+        help_text=_("Where this image came from."),
+    )
     categories = models.ManyToManyField(
         'icekit.MediaCategory',
         blank=True,
         related_name='%(app_label)s_%(class)s_related',
     )
-    admin_notes = models.TextField(
+    license = models.TextField(
+        _('License/rights information'),
+        blank=True,
+    )
+    notes = models.TextField(
         blank=True,
         help_text=_('Internal notes for administrators only.'),
     )
-    is_active = models.BooleanField(
+
+    width = models.PositiveIntegerField(editable=False)
+    height = models.PositiveIntegerField(editable=False)
+
+    date_created = models.DateTimeField(
+        default=timezone.now,
+        editable=False
+    )
+    date_modified = models.DateTimeField(
+        auto_now=True,
+        editable=False
+    )
+
+    # Unused for now
+    is_ok_for_web = models.BooleanField(
+        _("OK for web"),
         default=True,
+    )
+
+    # Unused for now
+    maximum_dimension = models.PositiveIntegerField(
+        blank=True, null=True,
+        help_text=_("If the size of this image is to be limited to a particular size for distribution, note it here."),
     )
 
     def clean(self):
@@ -53,6 +95,21 @@ class AbstractImage(models.Model):
 
     def __str__(self):
         return self.title or self.alt_text
+
+    def dimensions(self):
+        if self.width and self.height:
+            # NB using the multiplication symbol ×, not the letter x
+            return u"%s × %s" % (self.width, self.heigh)
+        return None
+
+    def file_size(self):
+        """
+        Obtain the file size for the file in human readable format.
+
+        :return: String of file size with unit.
+        """
+        return filesizeformat(self.image.size)
+
 
 
 @python_2_unicode_compatible
@@ -71,6 +128,7 @@ class ImageLinkMixin(models.Model):
 
     title_override = models.CharField(max_length=512, blank=True)
     caption_override = models.TextField(blank=True)
+    # not allowing credit override yet
 
     caption_template = "icekit/plugins/image/_caption.html"
 
@@ -144,9 +202,21 @@ class ImageLinkMixin(models.Model):
         """
         self.title_override = ''
 
+    @property
+    def credit(self):
+        """
+        Obtains the credit override or the actual image title.
+
+        :return: Title text (safe).
+        """
+        if self.show_title:
+            return mark_safe(self.image.credit)
+        return None
+
     def displayed_caption(self):
         c = Context({'instance': self})
         return render_to_string(self.caption_template, c)
+
 
 @python_2_unicode_compatible
 class AbstractImageItem(ContentItem, ImageLinkMixin):