+++ docs/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 In development
 --------------
 
+-  Fleshed out Image model with more metadata and fields for rights/usage.
+
 -  Add ``icekit.workflow`` application to associate, manage, and filter
    workflow state information like status and user-assigment for
    arbitrary models.